Started new plan
Want to do the whole thing in one go, but afraid to do too much in one commit.
New plan is:
- List of games
- Whenever someone joins, find an open game, give them that
So we can have one screen to show all, and there is less backlog if one person gets all the games. Also the game is less precariously placed on randomicity.